refactor: split between controllers and rest controllers
